requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
google-cloud-speech
google-cloud-texttospeech
langchain
chromadb
pypdf2
python-multipart
